requirements.txt
streamlit>=1.35.0
chromadb>=0.5.0
pypdf>=4.0.0
groq>=0.9.0
sentence-transformers>=2.7.0
langchain-text-splitters>=0.2.0
langchain-experimental>=0.0.59
langchain-huggingface>=0.0.3
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.26.4
protobuf>=4.25.0
tiktoken>=0.7.0